+++ critiqor/core_engine_dashboard/.output/server/index.mjs
@@ -85,108 +85,108 @@ const headers = ((m) => function headersRouteRule(event) {
   }
 });
 const assets = {
-  "/assets/card-3trT3p6x.js": {
+  "/assets/appearance-BSKLHDkp.js": {
     "type": "text/javascript; charset=utf-8",
-    "etag": '"b88-OwR5MNzOKggbit3cpok2YoaljrQ"',
-    "mtime": "2026-09-01T08:22:49.928Z",
-    "size": 2952,
-    "path": "../public/assets/card-3trT3p6x.js"
-  },
-  "/assets/checkbox-C_OZ0CGD.js": {
-    "type": "text/javascript; charset=utf-8",
-    "etag": '"2b2-N6sf84hVN3h4bTIcq5FYJb6y5v8"',
-    "mtime": "2026-09-01T08:22:49.928Z",
-    "size": 690,
-    "path": "../public/assets/checkbox-C_OZ0CGD.js"
-  },
-  "/assets/evidence-CdSLynlt.js": {
-    "type": "text/javascript; charset=utf-8",
-    "etag": '"135e-5R/8BEqJMCd85nd2FvmMweXmn08"',
-    "mtime": "2026-09-01T08:22:49.928Z",
-    "size": 4958,
-    "path": "../public/assets/evidence-CdSLynlt.js"
-  },
-  "/assets/diagnoses-m_npymSi.js": {
-    "type": "text/javascript; charset=utf-8",
-    "etag": '"94e7-vF5yQXkwwjnZefIE9Ec++p2OPRs"',
-    "mtime": "2026-09-01T08:22:49.928Z",
-    "size": 38119,
-    "path": "../public/assets/diagnoses-m_npymSi.js"
-  },
-  "/assets/playbook-BpV12O2I.js": {
-    "type": "text/javascript; charset=utf-8",
-    "etag": '"144f-85RJBDCvZ6C7tnD8xDpC6nLIQX0"',
-    "mtime": "2026-09-01T08:22:49.928Z",
-    "size": 5199,
-    "path": "../public/assets/playbook-BpV12O2I.js"
-  },
-  "/assets/run-selector-D0ijji2g.js": {
-    "type": "text/javascript; charset=utf-8",
-    "etag": '"38a-HjyZImk+Ig8FlWXAJFvxk7Dlfyw"',
-    "mtime": "2026-09-01T08:22:49.928Z",
-    "size": 906,
-    "path": "../public/assets/run-selector-D0ijji2g.js"
-  },
-  "/assets/runs-BUPkdfhf.js": {
-    "type": "text/javascript; charset=utf-8",
-    "etag": '"19da-3bt4FteanPlt4qF7YAR/YMIxJW0"',
-    "mtime": "2026-09-01T08:22:49.928Z",
-    "size": 6618,
-    "path": "../public/assets/runs-BUPkdfhf.js"
-  },
-  "/assets/appearance-gwAL_ked.js": {
-    "type": "text/javascript; charset=utf-8",
-    "etag": '"53a-lbGgHMLmngLGdgomiMrWfU0HBw0"',
-    "mtime": "2026-09-01T08:22:49.928Z",
+    "etag": '"53a-LQ60x5HTjIOXYZ1Uz5sc7nhuGIM"',
+    "mtime": "2026-09-03T09:08:48.567Z",
     "size": 1338,
-    "path": "../public/assets/appearance-gwAL_ked.js"
+    "path": "../public/assets/appearance-BSKLHDkp.js"
+  },
+  "/assets/card-DV18getr.js": {
+    "type": "text/javascript; charset=utf-8",
+    "etag": '"b88-IRtj10nKlI0DGqOSfPE+UPi67Cg"',
+    "mtime": "2026-09-03T09:08:48.567Z",
+    "size": 2952,
+    "path": "../public/assets/card-DV18getr.js"
+  },
+  "/assets/checkbox-CDVFq07J.js": {
+    "type": "text/javascript; charset=utf-8",
+    "etag": '"2b2-r8pUiXkfTqlf/CeeNpsLobx+els"',
+    "mtime": "2026-09-03T09:08:48.567Z",
+    "size": 690,
+    "path": "../public/assets/checkbox-CDVFq07J.js"
+  },
+  "/assets/diagnoses-C-XWYpNU.js": {
+    "type": "text/javascript; charset=utf-8",
+    "etag": '"94e7-WB1NCvhqXlI2ufyzVckchbOagHs"',
+    "mtime": "2026-09-03T09:08:48.566Z",
+    "size": 38119,
+    "path": "../public/assets/diagnoses-C-XWYpNU.js"
+  },
+  "/assets/evidence-DAqEdL6I.js": {
+    "type": "text/javascript; charset=utf-8",
+    "etag": '"135e-fXxiAZL9gjNR1wABmTZrqCRA+5A"',
+    "mtime": "2026-09-03T09:08:48.566Z",
+    "size": 4958,
+    "path": "../public/assets/evidence-DAqEdL6I.js"
+  },
+  "/assets/index-Clg_Wv4R.js": {
+    "type": "text/javascript; charset=utf-8",
+    "etag": '"228e-FCWKLCOmYeqxJ/mETug55mK4Wdc"',
+    "mtime": "2026-09-03T09:08:48.567Z",
+    "size": 8846,
+    "path": "../public/assets/index-Clg_Wv4R.js"
+  },
+  "/assets/playbook-BWd0ZG4H.js": {
+    "type": "text/javascript; charset=utf-8",
+    "etag": '"144f-DFc5WQfM5UoCF3YkSJ1dWN6Wg0Q"',
+    "mtime": "2026-09-03T09:08:48.567Z",
+    "size": 5199,
+    "path": "../public/assets/playbook-BWd0ZG4H.js"
+  },
+  "/assets/run-selector-DEi9lmHA.js": {
+    "type": "text/javascript; charset=utf-8",
+    "etag": '"38a-9IPdY0knxL0Md3qmc7AAdjnV5kg"',
+    "mtime": "2026-09-03T09:08:48.567Z",
+    "size": 906,
+    "path": "../public/assets/run-selector-DEi9lmHA.js"
+  },
+  "/assets/settings-DMywAjaX.js": {
+    "type": "text/javascript; charset=utf-8",
+    "etag": '"a66-CQbChA0FREiV97NMhpqhE8SiAnk"',
+    "mtime": "2026-09-03T09:08:48.566Z",
+    "size": 2662,
+    "path": "../public/assets/settings-DMywAjaX.js"
+  },
+  "/assets/runs-D-MOATdj.js": {
+    "type": "text/javascript; charset=utf-8",
+    "etag": '"19da-wey0H65JaGLJQKcvLm5LWlT0f6g"',
+    "mtime": "2026-09-03T09:08:48.566Z",
+    "size": 6618,
+    "path": "../public/assets/runs-D-MOATdj.js"
   },
   "/assets/react-vendor-D7KJJ4pD.js": {
     "type": "text/javascript; charset=utf-8",
     "etag": '"31ef8-pI8JeMnNSBfOEV8Pu5MCJGcBAV8"',
-    "mtime": "2026-09-01T08:22:49.928Z",
+    "mtime": "2026-09-03T09:08:48.566Z",
     "size": 204536,
     "path": "../public/assets/react-vendor-D7KJJ4pD.js"
-  },
-  "/assets/ui-vendor-B33jwOqA.js": {
-    "type": "text/javascript; charset=utf-8",
-    "etag": '"f66a-s6ka4P5HibFqyB0H9y07rjSzABQ"',
-    "mtime": "2026-09-01T08:22:49.928Z",
-    "size": 63082,
-    "path": "../public/assets/ui-vendor-B33jwOqA.js"
   },
   "/assets/styles-CPOj_Ctq.css": {
     "type": "text/css; charset=utf-8",
     "etag": '"17aaa-WlKSDMCp58hnV7+IOSb8C5Sq3tA"',
-    "mtime": "2026-09-01T08:22:49.928Z",
+    "mtime": "2026-09-03T09:08:48.566Z",
     "size": 96938,
     "path": "../public/assets/styles-CPOj_Ctq.css"
   },
-  "/assets/index-C6SLmSYf.js": {
+  "/assets/index-SanaLvIj.js": {
     "type": "text/javascript; charset=utf-8",
-    "etag": '"228e-7lOrptft0B0rC/nFJDLj1EdHY6U"',
-    "mtime": "2026-09-01T08:22:49.928Z",
-    "size": 8846,
-    "path": "../public/assets/index-C6SLmSYf.js"
-  },
-  "/assets/index-O-EJdERA.js": {
-    "type": "text/javascript; charset=utf-8",
-    "etag": '"469b7-RMRs6Vg4o14dZ7VysiYLv194UOE"',
-    "mtime": "2026-09-01T08:22:49.928Z",
+    "etag": '"469b7-+cMPtksjvQ+SBU/9BvNBEb5v3RE"',
+    "mtime": "2026-09-03T09:08:48.566Z",
     "size": 289207,
-    "path": "../public/assets/index-O-EJdERA.js"
+    "path": "../public/assets/index-SanaLvIj.js"
   },
-  "/assets/settings-DcB4R2iF.js": {
+  "/assets/ui-vendor-B33jwOqA.js": {
     "type": "text/javascript; charset=utf-8",
-    "etag": '"a66-jdU0DUC4Bk1eEsoweHj1tUld6vo"',
-    "mtime": "2026-09-01T08:22:49.928Z",
-    "size": 2662,
-    "path": "../public/assets/settings-DcB4R2iF.js"
+    "etag": '"f66a-s6ka4P5HibFqyB0H9y07rjSzABQ"',
+    "mtime": "2026-09-03T09:08:48.566Z",
+    "size": 63082,
+    "path": "../public/assets/ui-vendor-B33jwOqA.js"
   },
   "/assets/critiqor-logo-v-ZKSU2V.png": {
     "type": "image/png",
     "etag": '"af113-z31pw/7m150Murb+Un/bCcfK6tE"',
-    "mtime": "2026-09-01T08:22:49.928Z",
+    "mtime": "2026-09-03T09:08:48.567Z",
     "size": 717075,
     "path": "../public/assets/critiqor-logo-v-ZKSU2V.png"
   }